#544 Split CompActDefBuilderSpecs file into 3 files
