+++ dsl/src/test/groovy/org/cristalise/dsl/test/lifecycle/definition/CompActDefBuilderSpecs.groovy
@@ -193,237 +193,4 @@ class CompActDefBuilderSpecs extends Specification implements CristalTestSetup {
         ea1Slot.getProperties().get('booleanVal') == true
         ea1Slot.getProperties().get('intVal') == 0
     }
-
-    def 'CompositeActivityDef can start with LoopDef'() {
-        when:
-        def ea1 = new ActivityDef('EA1', 0)
-    
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-StartLoop', version: 0) {
-            Layout {
-                Loop { Act(ea1) }
-            }
-        }
-
-        def loopDef = caDef.getChildren().find { it instanceof LoopDef }
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-StartLoop'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 4
-        caDef.childrenGraphModel.startVertex.class.simpleName == 'JoinDef'
-
-        loopDef
-        loopDef.properties.RoutingExpr == 'true'
-        loopDef.getOutGraphables().findAll { GraphableVertex v ->
-            v.getBuiltInProperty(PAIRING_ID) == loopDef.getBuiltInProperty(PAIRING_ID) }.size() == 1
-
-        loopDef.getInGraphables().collect { it.name } == ['EA1']
-        loopDef.getOutGraphables().collect { it.class.simpleName } == ['JoinDef', 'JoinDef']
-    }
-
-    def 'CompositeActivityDef can build a sequence including LoopDef'() {
-        when:
-        def first   = new ActivityDef('first', 0)
-        def looping = new ActivityDef('looping', 0)
-        def last    = new ActivityDef('last', 0)
-    
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-IncludeLoop', version: 0) {
-            Layout {
-                Act(first)
-                Loop {
-                    Act(looping)
-                }
-                Act(last)
-            }
-        }
-
-        def loopDef = caDef.getChildren().find { it instanceof LoopDef }
-        def lastDef = caDef.getChildren().find { it.name == 'last'}
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-IncludeLoop'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 6
-        caDef.childrenGraphModel.startVertex.name == "first"
-        ((GraphableVertex)(caDef.childrenGraphModel.startVertex)).getOutGraphables().collect { it.class.simpleName } == ['JoinDef']
-
-        loopDef
-        loopDef.getOutGraphables().findAll { GraphableVertex v ->
-            v.getBuiltInProperty(PAIRING_ID) == loopDef.getBuiltInProperty(PAIRING_ID) }.size() == 1
-
-        loopDef.getOutGraphables().collect { it.class.simpleName } == ['JoinDef', 'JoinDef']
-        loopDef.getInGraphables().collect { it.name } == ['looping']
-
-        lastDef
-        lastDef.getInGraphables().collect { it.class.simpleName } == ['JoinDef']
-        lastDef.getOutGraphables().size() == 0
-    }
-
-    def 'LoopDef can define RoutingScript'() {
-        when:
-        def looping = new ActivityDef('looping', 0)
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-Loop-RoutingScript', version: 0) {
-            Layout {
-                Loop(javascript: true) {
-                    Property(toto: 123)
-                    Act(looping)
-                }
-            }
-        }
-
-        def loopDef = caDef.getChildren().find { it instanceof LoopDef }
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-Loop-RoutingScript'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 4
-
-        loopDef
-        loopDef.properties.RoutingScriptName == 'javascript:true;'
-        loopDef.properties.RoutingScriptVersion == null;
-        loopDef.properties.toto == 123
-    }
-
-    def 'LoopDef can by Infinitive'() {
-        when:
-        def looping = new ActivityDef('looping', 0)
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-Loop-Infinitive', version: 0) {
-            Layout {
-                LoopInfinitive() {
-                    Property(toto: 123)
-                    Act(looping)
-                }
-            }
-        }
-
-        def loopDef = caDef.getChildren().find { it instanceof LoopDef }
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-Loop-Infinitive'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 4
-
-        loopDef
-        loopDef.properties.RoutingScriptName == 'groovy:true;'
-        loopDef.properties.RoutingScriptVersion == null;
-        loopDef.properties.toto == 123
-    }
-
-    def 'CompositeActivityDef can start and finish with AndSplit'() {
-        when:
-        def left  = new ActivityDef('left',  0)
-        def right = new ActivityDef('right', 0)
-
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-StartAndSplit', version: 0) {
-            Layout {
-                AndSplit {
-                    Block { Act(left)  }
-                    Block { Act(right) }
-                }
-            }
-        }
-
-        def andSplitDef = caDef.getChildren().find { it instanceof AndSplitDef }
-        def joinDef = caDef.getChildren().find { it instanceof JoinDef }
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-StartAndSplit'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 4
-        caDef.childrenGraphModel.startVertex.class.simpleName == 'AndSplitDef'
-
-        andSplitDef
-        joinDef
-
-        andSplitDef.getBuiltInProperty(PAIRING_ID)
-        joinDef.getBuiltInProperty(PAIRING_ID)
-        andSplitDef.getBuiltInProperty(PAIRING_ID) == joinDef.getBuiltInProperty(PAIRING_ID)
-
-        andSplitDef.getInGraphables().size() == 0
-        andSplitDef.getOutGraphables().collect {it.name} == ['left','right']
-        joinDef.getInGraphables().collect {it.name} == ['left','right']
-        joinDef.getOutGraphables().size() == 0
-    }
-
-    def 'CompositeActivityDef can build a sequence including AndSplit'() {
-        when:
-        def first = new ActivityDef('first', 0)
-        def left1 = new ActivityDef('left1', 0)
-        def left2 = new ActivityDef('left2', 0)
-        def right = new ActivityDef('right', 0)
-        def last  = new ActivityDef('last',  0)
-
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-IncludeAndSplit', version: 0) {
-            Layout {
-                Act(first)
-                AndSplit {
-                    Block {
-                        Act(left1)
-                        Act(left2)
-                    }
-                    Block {
-                        Act(right)
-                    }
-                }
-                Act(last)
-            }
-        }
-
-        def andSplitDef = caDef.getChildren().find { it instanceof AndSplitDef }
-        def joinDef = caDef.getChildren().find { it instanceof JoinDef }
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-IncludeAndSplit'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 7
-        caDef.childrenGraphModel.startVertex.name == 'first'
-
-        andSplitDef
-        joinDef
-
-        andSplitDef.getBuiltInProperty(PAIRING_ID)
-        joinDef.getBuiltInProperty(PAIRING_ID)
-        andSplitDef.getBuiltInProperty(PAIRING_ID) == joinDef.getBuiltInProperty(PAIRING_ID)
-
-        andSplitDef.getInGraphables().collect {it.name} == ['first']
-        andSplitDef.getOutGraphables().collect {it.name} == ['left1','right']
-        joinDef.getInGraphables().collect {it.name} == ['left2','right']
-        joinDef.getOutGraphables().collect {it.name} == ['last']
-    }
-
-    def 'CompositeActivityDef can define AndSplit with Loops'() {
-        when:
-        def left   = new ActivityDef('left',  0)
-        def middle = new ActivityDef('middle', 0)
-        def right  = new ActivityDef('right', 0)
-
-        caDef = CompActDefBuilder.build(module: 'test', name: 'CADef-AndSplitWithLoops', version: 0) {
-            Layout {
-                AndSplit {
-                    LoopInfinitive { Act('Left', left)  }
-                    Block { Act('Middle', middle) }
-                    Loop { Act('Right', right) }
-                    Block { CompActDef('ManageItemDesc', 0) }
-                }
-            }
-        }
-
-        def andSplitDef = (AndSplitDef) caDef.childrenGraphModel.startVertex
-
-        then:
-        caDef.verify()
-        caDef.name == 'CADef-AndSplitWithLoops'
-        caDef.version == 0
-        caDef.childrenGraphModel.vertices.length == 12
-        caDef.childrenGraphModel.startVertex.class.simpleName == 'AndSplitDef'
-
-        andSplitDef.getInGraphables().size() == 0
-        andSplitDef.getOutGraphables().collect { it.class.simpleName } == ['JoinDef', 'ActivitySlotDef', 'JoinDef', 'ActivitySlotDef']
-    }
 }
